+++ ComfyUI/custom_nodes/orpheus_tts/requirements.txt
@@ -1,1 +1,1 @@
-morpheus_tts
+Morpheus_Client
