+++ GOALS.md
@@ -135,7 +135,7 @@ _(Append new capabilities below using the format above. Keep the list curated; c
 
 ### Capability: morpheus-client-introspection
 - **Purpose:** Allow clients to discover available voices and adapter capabilities.
-- **Scope:** `Morpheus_Client/api/server.py`, `INTERFACES.md`
+- **Scope:** `Morpheus_Client/server.py`, `INTERFACES.md`
 - **Shape:** `GET /v1/audio/voices` → `{status, voices}`; `GET /adapters` → `{adapter_name: descriptor}`
 - **Compatibility:** read-only; no flags
 - **Status:** active
