+++ Morpheus_Client/orchestrator/stitcher.py
